+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Next version
 
 - Handle the case when a page's site and its parent's site do not match.
 - Added Python 3.12, Django 5.0.
+- Dropped support for Django before 4.2.
 - Changed the ``DefaultLanguageListFilter`` to use the model admin's default
   empty value display, and allowed filtering for sites where a default value
   isn't set.